Enhanced Script Testing
- I got a little paranoid around the SVScript's robustness after the
  last bug report. Adding more comprehensive testing to the SVScript
class.
- Started framing the tests for the LockBuilder/UnlockBuilder classes
  for pending transaction templates.
