+++ test/transaction/p2pk_builder_test.dart
@@ -1,0 +1,16 @@
+
+void main() {
+
+}
+/*
+
+  describe('#buildPublicKeyOut', function () {
+    it('should create script from public key', function () {
+      var pubkey = new PublicKey('022df8750480ad5b26950b25c7ba79d3e37d75f640f8e5d9bcd5b150a0f85014da')
+      var s = Script.buildPublicKeyOut(pubkey)
+      should.exist(s)
+      s.toString().should.equal('33 0x022df8750480ad5b26950b25c7ba79d3e37d75f640f8e5d9bcd5b150a0f85014da OP_CHECKSIG')
+      s.isPublicKeyOut().should.equal(true)
+    })
+  })
+ */
